requirement.txt
requests
Pillow
PyPDF2
